+++ apps/guardian/src-tauri/src/config.rs
@@ -1,10 +1,10 @@
 use std::{str::FromStr, sync::LazyLock};
 
-use anyhow::{anyhow, Context};
+use anyhow::Context;
 use concordium_governance_committee_election::{self as contract};
 use concordium_rust_sdk::{
     common::types::Amount,
-    contract_client::{ContractClient, ContractUpdateBuilder},
+    contract_client::ContractUpdateBuilder,
     id::types::AccountAddress,
     v2::{self, BlockIdentifier},
 };
@@ -13,6 +13,7 @@ use eg::{
     guardian_public_key::GuardianPublicKey, guardian_share::GuardianEncryptedShare,
 };
 use election_common::{
+    contract::{verify_contract, ElectionClient},
     decode, encode, EncryptedTally, GuardianDecryption, GuardianDecryptionProof, HttpClient,
 };
 use tonic::transport::ClientTlsConfig;
@@ -31,18 +32,6 @@ static TIMEOUT: LazyLock<u64> = LazyLock::new(|| {
 static HTTP_CLIENT: LazyLock<HttpClient> =
     LazyLock::new(|| HttpClient::try_create(*TIMEOUT).expect("Failed to create HTTP client"));
 
-/// The necessary election guard configuration to construct election guard
-/// entities.
-#[derive(Clone)]
-pub struct ElectionGuardConfig {
-    /// The election manifest
-    pub manifest:   ElectionManifest,
-    /// The election parameters
-    pub parameters: ElectionParameters,
-}
-
-pub struct ElectionContractMarker;
-pub type ElectionClient = ContractClient<ElectionContractMarker>;
 /// The election contract client
 #[derive(Clone)]
 pub struct ElectionContract(pub ElectionClient);
@@ -56,7 +45,8 @@ impl ElectionContract {
                 &(),
                 BlockIdentifier::LastFinal,
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to get config from contract: {e}"))?;
 
         Ok(config)
     }
@@ -69,7 +59,8 @@ impl ElectionContract {
                 &(),
                 BlockIdentifier::LastFinal,
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to get encrypted tally from contract: {e}"))?;
         let Some(tally) = tally else {
             return Ok(None);
         };
@@ -88,7 +79,8 @@ impl ElectionContract {
                 &(),
                 BlockIdentifier::LastFinal,
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to get guardians state from contract: {e}"))?;
 
         Ok(state)
     }
@@ -106,7 +98,8 @@ impl ElectionContract {
                 *sender,
                 &encode(public_key).unwrap(), // Serialization will not fail
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to register public key in contract: {e}"))?;
 
         Ok(update)
     }
@@ -124,7 +117,8 @@ impl ElectionContract {
                 *sender,
                 &encode(shares).unwrap(), // Serialization will not fail
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to register encrypted share in contract: {e}"))?;
 
         Ok(update)
     }
@@ -142,7 +136,8 @@ impl ElectionContract {
                 *sender,
                 guardian_status,
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to register guardian status in contract: {e}"))?;
 
         Ok(update)
     }
@@ -160,7 +155,8 @@ impl ElectionContract {
                 *sender,
                 &encode(decryption).unwrap(), // Serialization will not fail
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to post decryption in contract: {e}"))?;
 
         Ok(update)
     }
@@ -178,10 +174,21 @@ impl ElectionContract {
                 *sender,
                 &encode(shares).unwrap(), // Serialization will not fail
             )
-            .await?;
+            .await
+            .inspect_err(|e| log::error!("Failed to post decryption proof in contract: {e}"))?;
 
         Ok(update)
     }
+}
+
+/// The necessary election guard configuration to construct election guard
+/// entities.
+#[derive(Clone)]
+pub struct ElectionGuardConfig {
+    /// The election manifest
+    pub manifest:   ElectionManifest,
+    /// The election parameters
+    pub parameters: ElectionParameters,
 }
 
 pub struct AppConfig {
@@ -248,14 +255,16 @@ impl AppConfig {
         let genesis_hash = node.get_consensus_info().await?.genesis_block;
         let expected_genesis_hash = network.genesis_hash();
         if genesis_hash != expected_genesis_hash {
-            return Err(anyhow!(
-                "Invalid node specified. Application must use a {} node",
+            return Err(Error::InvalidConfiguration(format!(
+                "Wrong node configuration, expected a {} node",
                 network
-            )
-            .into());
+            )));
         }
 
-        let contract = ElectionContract(ElectionClient::create(node, contract_address).await?);
+        let contract = verify_contract(node, contract_address).await.map_err(|_| {
+            Error::InvalidConfiguration("Failed to verify election contract".to_string())
+        })?;
+        let contract = ElectionContract(contract);
         self.contract = Some(contract.clone());
         Ok(contract)
     }