+++ apps/guardian/CHANGELOG.md
@@ -10,7 +10,8 @@
 
 - Bumped rust-sdk dependency to 6.0
 - Update `@concordium/web-sdk` to 9.1.
-- Added configuration by user config file, allowing users to configure the network/contract to integrate with
+- Added configuration by user config file, allowing users to configure the network/contract to integrate with. This can
+  be accessed from the "Settings" menu from the native application menu
 
 
 ## 1.0.0